spec: create auth helper
