+++ spec/requests/sessions_spec.rb
@@ -88,17 +88,4 @@ RSpec.describe "Sessions" do
       expect(flash[:notice]).to eq("That session has been logged out")
     end
   end
-
-  # TODO: Move to a request helper
-  def sign_in(user = nil)
-    user ||= create(:user)
-
-    # rubocop:disable RSpec/AnyInstance
-    allow_any_instance_of(Current)
-      .to receive(:session).and_return(user.sessions.create!)
-
-    allow_any_instance_of(ApplicationController)
-      .to receive(:authenticate).and_return(true)
-    # rubocop:enable RSpec/AnyInstance
-  end
 end
